refactor viability calculation to use BigDecimal for cost values and add CORS configuration

+++ api/src/main/java/br/com/viability/api/viability/dto/ViabilityResponseDto.java
@@ -1,7 +1,9 @@
 package br.com.viability.api.viability.dto;
 
+import java.math.BigDecimal;
+
 public record ViabilityResponseDto(
-        Double totalCost,
-        Double costPerM2
+        BigDecimal totalCost,
+        BigDecimal costPerM2
 ) {
 }
